separate ms_exec
